(PC-18978) feat(offerDuo): Remove filter OfferType, and make it work only for OfferDuo (wording as Duo)
